update na App.java

+++ src/main/java/ba/unsa/etf/rpr/Faktorijel.java
@@ -1,7 +1,7 @@
 package ba.unsa.etf.rpr;
 
 public class Faktorijel {
-    public static double izracunaj(int broj) {
+    public static double izracunaj(double broj) {
         double fakt=1;
         while(broj>1){
             fakt = fakt * broj;
