+++ src/main/java/ba/unsa/etf/rpr/Sinus.java
@@ -3,7 +3,7 @@ package ba.unsa.etf.rpr;
 import static java.lang.Math.sin;
 
 public class Sinus {
-    public static double izracunaj(int broj) {
+    public static double izracunaj(double broj) {
         return sin(broj);
     }
 }
